Adding userlist message type to client and server.
The point of the userlist is to allow a client to bootstrap it's list of
users by simply asking the server for all the users in the room.  This way
multiple clients connecting to the server at different times will still
see the same user list.  Of course, the order is different on each client
so the next step is to get some basic sorting working in the handle_join
and handle_userlist methods.